+++ app/Http/Controllers/API/GameController.php
@@ -56,70 +56,6 @@ class GameController extends Controller
         }
     }
 
-    public function showAllGamesFromUser($id)
-    {
-        $user_auth_id = Auth::id();
-        $user_auth = User::find($user_auth_id);
-        $user = User::find($id);
-
-        if (!$user) {
-            return response([
-                'message' => 'Unregistred User',
-                'status' => 404,
-            ]);
-        } elseif ($user_auth_id == $id || $user_auth->hasRole('Admin')) {
-            $games = $user->games;
-            $countGames = count($games);
-            if (!$countGames) {
-                return response([
-                    'message' => 'The user ' . $user->name . ' has not played yet.',
-                    'status' => 200,
-                ]);
-            } else {
-                return response([
-                    'games' => $games,
-                    'message' => 'Retrieved Succesfully',
-                    'status' => 200,
-                ]);
-            }
-        } else {
-            return response([
-                'message' => 'Sorry, you are not authorized to perform this action.',
-                'status' => 401,
-            ]);
-        }
-    }
-
-    public function destroyThrowsFromUser($id)
-    {
-        // A specific game cannot be deleted. 
-        // Only a user's total execution log can be deleted.
-        $user_auth_id = Auth::id();
-        $user_auth = User::find($user_auth_id);
-        $user = User::find($id);
-
-        if (!$user) {
-            return response([
-                'message' => 'Unregistred User',
-                'status' => 404,
-            ]);
-        } elseif ($user_auth_id == $id || $user_auth->hasRole('Admin')) {
-            $allGamesFromUser = $user->games();
-            $allGamesFromUser->delete();
-
-            return response([
-                'game' => 0,
-                'message' => 'All Data Deleted From The User ' . ucfirst($user->name),
-                'status' => 200,
-            ]);
-        } else {
-            return response([
-                'message' => 'Sorry, you are not authorized to perform this action.',
-                'status' => 401,
-            ]);
-        }
-    }
-
     public function fullSuccessRateRecord()
     {
         $fullSuccessRateRecord = DB::table('games')
